requirements.txt
asgiref==3.8.1
Brotli==1.1.0
click==8.1.7
dj-database-url==2.1.0
Django==4.2.11
django-cors-headers==4.4.0
django-extensions==3.2.3
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-jsonapi==6.1.0
gunicorn==22.0.0
h11==0.14.0
inflection==0.5.1
packaging==24.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
uvicorn==0.29.0
whitenoise==6.6.0
